Move DUP alert-fetching logic into separate module
